Split Graph Sketcher into multiple files
