+++ app/js/app/directives/graph_sketcher/GraphSketcher.js
@@ -1,43 +1,127 @@
 "use strict";
-define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketcher/linear.js", "../../../lib/graph_sketcher/func.js", "../../../lib/graph_sketcher/sampler.js", "/partials/graph_sketcher/graph_sketcher.html"],
-    function(p5, bezierLineType, linearLineType, graphFunctions, lineSampler, templateUrl) {
+define(["p5", "./GraphView.js", "./GraphUtils.js", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketcher/linear.js", "/partials/graph_sketcher/graph_sketcher.html"],
+    function(p5, graphView, graphUtils, bezierLineType, linearLineType, templateUrl) {
     return ["$timeout", "$rootScope", "api", function($timeout, $rootScope, api) {
         let instanceCounter = 0;
         return {
-            // scope: true,
-            // transclude: true,
             restrict: "A",
             templateUrl: templateUrl,
 
             link: function(scope, element, attrs) {
 
-                element.on("touchstart touchmove", "canvas", function(e) {
-                    e.preventDefault();
-                });
-
                 scope.title = "Sketcher";
-                scope.canvasOffset = {};
-                scope.draggingNewSymbol = false;
-                scope.equationEditorElement = element;
                 scope.selectedLineType = bezierLineType;
 
+                // objects which are shared between sketch and the ui buttons
                 let colorSelect = element.find(".color-select")[0];
                 let encodeData;
                 let decodeData;
                 let reDraw;
+                let freeSymbols = [];
+                let curves = [];
+                let clickedKnot = null;
+                let clickedCurveIdx;
 
+                function initiateFreeSymbols() {
+                    freeSymbols = [];
 
+                    const symbolList = ['A', 'B', 'C', 'D', 'E'];
+                    symbolList.forEach(function(symbol) {
+                        freeSymbols.push(graphUtils.createSymbol(symbol));
+                    });
+                    
+                    refreshFreeSymbols();
+                }
 
-                scope.logOnClose = function(event) {
-                    // This ought to catch people who navigate away without closing the editor!
-                    if (scope.log != null) {
-                        scope.log.actions.push({
-                            event: "NAVIGATE_AWAY",
-                            timestamp: Date.now()
-                        });
-                        api.logger.log(scope.log);
+                function refreshFreeSymbols() {
+                    let start = 15,
+                        separation = 30;
+
+                    for (let i = 0; i < freeSymbols.length; i++) {
+                        let symbol = freeSymbols[i];
+                        symbol.x = start + i * separation;
+                        symbol.y = start;
                     }
+                }
+
+                element.on("touchstart touchmove", "canvas", function(e) {
+                    e.preventDefault();
+                });
+
+                scope.undo = function() {
+                    if (scope.p.checkPointsUndo.length == 0 || scope.p.checkPointsUndo === undefined) {
+                        return;
+                    }
+
+                    let checkPointRedo = {};
+                    checkPointRedo.freeSymbolsJSON = JSON.stringify(freeSymbols);
+                    checkPointRedo.curvesJSON = JSON.stringify(curves);
+                    scope.p.checkPointsRedo.push(checkPointRedo);
+
+                    let checkPointUndo = scope.p.checkPointsUndo.pop();
+                    freeSymbols = JSON.parse(checkPointUndo.freeSymbolsJSON);
+                    curves = JSON.parse(checkPointUndo.curvesJSON);
+                    
+                    clickedKnot = null;
+                    clickedCurveIdx = undefined;
+
+                    reDraw();
+                }
+
+                scope.redo = function() {
+                    event.stopPropagation();
+                    if (scope.p.checkPointsRedo.length == 0) {
+                        return;
+                    }
+
+                    let checkPointUndo = {};
+                    checkPointUndo.freeSymbolsJSON = JSON.stringify(freeSymbols);
+                    checkPointUndo.curvesJSON = JSON.stringify(curves);
+                    scope.p.checkPointsUndo.push(checkPointUndo);
+
+                    let checkPointRedo = scope.p.checkPointsRedo.pop();
+                    freeSymbols = JSON.parse(checkPointRedo.freeSymbolsJSON);
+                    curves = JSON.parse(checkPointRedo.curvesJSON);
+                    
+                    clickedKnot = null;
+                    clickedCurveIdx = undefined;
+                    reDraw();
+                }
+
+                scope.isUndoable = function() {
+                    return scope.p && scope.p.checkPointsUndo.length > 0;
+                }
+
+                scope.isRedoable = function() {
+                    return scope.p && scope.p.checkPointsRedo.length > 0;
+                }
+
+                scope.straight = function() {
+                    scope.selectedLineType = linearLineType;
+                }
+
+                scope.poly = function() {
+                    scope.selectedLineType = bezierLineType;
+                }
+
+                scope.clean = function() {
+                    scope.p.checkPoint = {};
+                    scope.p.checkPoint.freeSymbolsJSON = JSON.stringify(freeSymbols);
+                    scope.p.checkPoint.curvesJSON = JSON.stringify(curves);
+                    scope.p.checkPointsUndo.push(scope.p.checkPoint);
+                    scope.p.checkPointsRedo = [];
+
+                    curves = [];
+                    clickedKnot = null;
+                    clickedCurveIdx = undefined;
+                    initiateFreeSymbols();
+                    reDraw();
+                }
+
+                scope.submit = function() {
+                    $("#graphModal").foundation("reveal", "close");
                 };
+
 
                 scope.sketch = function(p) {
 
@@ -45,20 +129,9 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                     let canvasHeight = window.innerHeight;
                     let canvasWidth = window.innerWidth;
 
-                    let CURVE_LIMIT = 3;
-
-                    let GRID_WIDTH = 60;
-                    let CURVE_STRKWEIGHT = 2;
-                    let PADDING = 0.025 * canvasWidth;
-                    let DOT_LINE_STEP = 5;
-                    let MOUSE_DETECT_RADIUS = 10;
-
-                    let CURVE_COLORS = [[93,165,218], [250,164,58], [96,189,104], [241,124,176], [241,88,84], [178,118,178]];
-                    let KNOT_COLOR = [77,77,77];
-                    let DOT_LINE_COLOR = [123];
-                    let MOVE_LINE_COLOR = [135];
-                    let MOVE_SYMBOL_COLOR = [151];
-                    let KNOT_DETECT_COLOR = [0];
+                    const MOVE_SYMBOL_COLOR = [151];
+                    const CURVE_LIMIT = 3;
+                    const MOUSE_DETECT_RADIUS = 10;
 
                     // action recorder
                     let action = undefined;
@@ -66,9 +139,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                     let releasePt;
                     let drawMode;
                     let key = undefined;
-                        
-                    let freeSymbols = [];
-                    let curves = [];
+
                     let dat;
 
 
@@ -91,8 +162,6 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                     let bindedKnot;
                     let symbolType;
 
-                    let clickedKnot = null;
-                    let clickedCurveIdx;
                     let clickedKnotId;
                     let tempCurve;
                     let tempPts = [];
@@ -103,27 +172,6 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                     p.checkPointsUndo = [];
                     p.checkPointsRedo = [];
 
-                    function initiateFreeSymbols() {
-                        freeSymbols = [];
-
-                        const symbolList = ['A', 'B', 'C', 'D', 'E'];
-                        symbolList.forEach(function(symbol) {
-                            freeSymbols.push(graphFunctions.createSymbol(symbol));
-                        });
-                        
-                        refreshFreeSymbols();
-                    }
-
-                    function refreshFreeSymbols() {
-                        let start = 15,
-                            separation = 30;
-
-                        for (let i = 0; i < freeSymbols.length; i++) {
-                            let symbol = freeSymbols[i];
-                            symbol.x = start + i * separation;
-                            symbol.y = start;
-                        }
-                    }
 
                     // run in the beginning by p5 library
                     p.setup = function() {
@@ -139,7 +187,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                         let current = getMousePt(e);
 
 
-                        // this funciton does not react if the mouse is over buttons or outside the canvas.
+                        // this function does not react if the mouse is over buttons or outside the canvas.
                         if (!isActive(current)) {
                             return;
                         }
@@ -155,9 +203,9 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 }
 
                                 for (let i = 0; i < knots.length; i++) {
-                                    if (graphFunctions.getDist(current, knots[i]) < MOUSE_DETECT_RADIUS) {
+                                    if (graphUtils.getDist(current, knots[i]) < MOUSE_DETECT_RADIUS) {
                                         p.cursor(p.HAND);
-                                        drawKnotDetect(knots[i]);
+                                        graphView.drawKnotDetect(p, knots[i]);
                                         found = true;
                                         return;
                                     }
@@ -191,7 +239,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                         if (!found) {
                             for (let i = 0; i < curves.length; i++) {
                                 for (let j = 0; j < curves[i].pts.length; j++) {
-                                    if (graphFunctions.getDist(current, curves[i].pts[j]) < MOUSE_DETECT_RADIUS) {
+                                    if (graphUtils.getDist(current, curves[i].pts[j]) < MOUSE_DETECT_RADIUS) {
                                         p.cursor(p.MOVE);
                                         found = true;
                                         break;
@@ -276,7 +324,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     p.cursor(p.MOVE);
                                 } else if (detect(c.minX, c.minY) || detect(c.maxX, c.minY) || detect(c.minX, c.maxY) || detect(c.maxX, c.maxY)) {
                                     p.push();
-                                    p.fill(KNOT_DETECT_COLOR);
+                                    p.fill(graphView.KNOT_DETECT_COLOR);
                                     if (detect(c.minX, c.minY)) {
                                          p.rect(c.minX - 4, c.minY - 4, 8, 8);
                                     } else if (detect(c.maxX, c.minY)) {
@@ -294,7 +342,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     || detect(c.minX - 3, (c.minY + c.maxY)/2) || detect(c.maxX + 3, (c.minY + c.maxY)/2)) {
 
                                     p.push();
-                                    p.fill(KNOT_DETECT_COLOR);
+                                    p.fill(graphView.KNOT_DETECT_COLOR);
                                     if (detect((c.minX + c.maxX)/2, c.minY - 3)) {
                                         p.triangle((c.minX + c.maxX)/2 - 5, c.minY - 2, (c.minX + c.maxX)/2 + 5, c.minY - 2, (c.minX + c.maxX)/2, c.minY - 7);
                                     } else if (detect((c.minX + c.maxX)/2, c.maxY + 3)) {
@@ -339,7 +387,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                         let current = getMousePt(e);
                         releasePt = current;
 
-                        // this funciton does not react if the mouse is over buttons or outside the canvas.
+                        // this function does not react if the mouse is over buttons or outside the canvas.
                         if (!isActive(current)) {
                             return;
                         }
@@ -482,7 +530,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 let minima = curves[i].minima;
                                 for (let j = 0; j < maxima.length; j++) {
                                     let knot = maxima[j];
-                                    if (graphFunctions.getDist(current, knot) < MOUSE_DETECT_RADIUS + 10){
+                                    if (graphUtils.getDist(current, knot) < MOUSE_DETECT_RADIUS + 10){
                                         clickedCurve = i;
                                         action = "STRETCH_POINT";
                                         clickedKnotId = j;
@@ -494,7 +542,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 }
                                 for (let j = 0; j < minima.length; j++) {
                                     let knot = minima[j];
-                                    if (graphFunctions.getDist(current, knot) < MOUSE_DETECT_RADIUS + 10){
+                                    if (graphUtils.getDist(current, knot) < MOUSE_DETECT_RADIUS + 10){
                                         clickedCurve = i;
                                         action = "STRETCH_POINT";
                                         clickedKnotId = j;
@@ -508,7 +556,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             let tc = [];
                             for (let i = 0; i < curves.length; i++) {
                                 for (let j = 0; j < curves[i].pts.length; j++) {
-                                    if (graphFunctions.getDist(current, curves[i].pts[j]) < MOUSE_DETECT_RADIUS) {
+                                    if (graphUtils.getDist(current, curves[i].pts[j]) < MOUSE_DETECT_RADIUS) {
                                         clickedCurveIdx = i;
                                         tc = curves[clickedCurveIdx];
                                         break;
@@ -516,7 +564,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 }
                             } 
                             if (tc != undefined) {   
-                                  // && graphFunctions.getDist(current, knot) > MOUSE_DETECT_RADIUS + 10
+                                  // && graphUtils.getDist(current, knot) > MOUSE_DETECT_RADIUS + 10
                                 if (current.x >= tc.minX && current.x <= tc.maxX && current.y >= tc.minY && current.y <= tc.maxY) {
                                     movedCurveIdx = clickedCurveIdx;
                                     action = "MOVE_CURVE";
@@ -614,9 +662,9 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 for (i = 0; i < m.pts.length; i++) {
                                     m.pts[i].ind = i;
                                 }
-                                m.endPt = findEndPts(m.pts);
-                                m.maxima = findTurnPts(m.pts, 'maxima');
-                                m.minima = findTurnPts(m.pts, 'minima');
+                                m.endPt = graphUtils.findEndPts(m.pts);
+                                m.maxima = graphUtils.findTurnPts(m.pts, 'maxima');
+                                m.minima = graphUtils.findTurnPts(m.pts, 'minima');
 
                                 reDraw();
                             }
@@ -691,14 +739,14 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     stretchMode = 2;
                                     switch (stretchMode) {
                                         case 2: {
-                                            stretchCurve(tempCurveL, lorx, lory, lnrx, lnry, tempMin.x, tempMin.y);
+                                            graphUtils.stretchCurve(tempCurveL, lorx, lory, lnrx, lnry, tempMin.x, tempMin.y);
                                             break;
                                         }
                                     }
                                     stretchMode = 3;
                                     switch (stretchMode) {
                                         case 3: {
-                                            stretchCurve(tempCurveH, rorx, rory, rnrx, rnry, tempMax.x, tempMax.y);
+                                            graphUtils.stretchCurve(tempCurveH, rorx, rory, rnrx, rnry, tempMax.x, tempMax.y);
                                             break;
                                         }
                                     }
@@ -714,10 +762,10 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     m.pts = intpts;
                                     // m.pts = uniqueNames;
 
-                                    m.interX = findInterceptX(m.pts);
-                                    m.interY = findInterceptY(m.pts);     
-                                    m.maxima = findTurnPts(m.pts, 'maxima');   
-                                    m.minima = findTurnPts(m.pts, 'minima');  
+                                    m.interX = graphUtils.findInterceptX(canvasHeight, m.pts);
+                                    m.interY = graphUtils.findInterceptY(canvasWidth, m.pts);     
+                                    m.maxima = graphUtils.findTurnPts(m.pts, 'maxima');   
+                                    m.minima = graphUtils.findTurnPts(m.pts, 'minima');  
                                     let minX = m.pts[0].x;
                                     let maxX = m.pts[0].x;
                                     let minY = m.pts[0].y;
@@ -798,14 +846,14 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     stretchMode = 1;
                                     switch (stretchMode) {
                                         case 1: {
-                                            stretchCurve(tempCurveL, lorx, lory, lnrx, lnry, tempMin.x, tempMin.y);
+                                            graphUtils.stretchCurve(tempCurveL, lorx, lory, lnrx, lnry, tempMin.x, tempMin.y);
                                             break;
                                         }
                                     }
                                     stretchMode = 0;
                                     switch (stretchMode) {
                                         case 0: {
-                                            stretchCurve(tempCurveH, rorx, rory, rnrx, rnry, tempMax.x, tempMax.y);
+                                            graphUtils.stretchCurve(tempCurveH, rorx, rory, rnrx, rnry, tempMax.x, tempMax.y);
                                             break;
                                         }
                                     }
@@ -819,10 +867,10 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     intpts.push.apply(intpts, upperRng);
                                     m.pts = intpts;
                                     
-                                    m.interX = findInterceptX(m.pts);
-                                    m.interY = findInterceptY(m.pts);     
-                                    m.maxima = findTurnPts(m.pts, 'maxima');   
-                                    m.minima = findTurnPts(m.pts, 'minima');
+                                    m.interX = graphUtils.findInterceptX(canvasHeight, m.pts);
+                                    m.interY = graphUtils.findInterceptY(canvasWidth, m.pts);     
+                                    m.maxima = graphUtils.findTurnPts(m.pts, 'maxima');   
+                                    m.minima = graphUtils.findTurnPts(m.pts, 'minima');
                                     let minX = m.pts[0].x;
                                     let maxX = m.pts[0].x;
                                     let minY = m.pts[0].y;
@@ -851,16 +899,16 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             let dx = current.x - prevMousePt.x;
                             let dy = current.y - prevMousePt.y;
                             prevMousePt = current;
-                            transCurve(curves[movedCurveIdx], dx, dy);
+                            graphUtils.transCurve(curves[movedCurveIdx], dx, dy);
 
                             reDraw();
-                            drawCurve(curves[movedCurveIdx], MOVE_LINE_COLOR);
+                            graphView.drawCurve(p, curves[movedCurveIdx], MOVE_LINE_COLOR);
 
                         } else if (action == "STRETCH_CURVE") {
 
                             function drawCorner() {
                                 p.push();
-                                p.fill(KNOT_DETECT_COLOR);
+                                p.fill(graphView.KNOT_DETECT_COLOR);
                                 switch (stretchMode) {
                                     case 0: {
                                         p.rect(c.minX - 4, c.minY - 4, 8, 8);
@@ -983,35 +1031,35 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             // stretch the curve
                             switch (stretchMode) {
                                 case 0: {
-                                    stretchCurve(c, orx, ory, nrx, nry, c.maxX, c.maxY);
+                                    graphUtils.stretchCurve(c, orx, ory, nrx, nry, c.maxX, c.maxY);
                                     break;
                                 }
                                 case 1: {
-                                    stretchCurve(c, orx, ory, nrx, nry, c.minX, c.maxY);
+                                    graphUtils.stretchCurve(c, orx, ory, nrx, nry, c.minX, c.maxY);
                                     break;
                                 }
                                 case 2: {
-                                    stretchCurve(c, orx, ory, nrx, nry, c.minX, c.minY);
+                                    graphUtils.stretchCurve(c, orx, ory, nrx, nry, c.minX, c.minY);
                                     break;
                                 }
                                 case 3: {
-                                    stretchCurve(c, orx, ory, nrx, nry, c.maxX, c.minY);
+                                    graphUtils.stretchCurve(c, orx, ory, nrx, nry, c.maxX, c.minY);
                                     break;
                                 }
                                 case 4: {
-                                    stretchCurve(c, orx, ory, orx, nry, (c.minX + c.maxX)/2, c.maxY);
+                                    graphUtils.stretchCurve(c, orx, ory, orx, nry, (c.minX + c.maxX)/2, c.maxY);
                                     break;
                                 }
                                 case 5: {
-                                    stretchCurve(c, orx, ory, orx, nry, (c.minX + c.maxX)/2, c.minY);
+                                    graphUtils.stretchCurve(c, orx, ory, orx, nry, (c.minX + c.maxX)/2, c.minY);
                                     break;
                                 }
                                 case 6: {
-                                    stretchCurve(c, orx, ory, nrx, ory, c.maxX, (c.minY + c.maxY)/2);
+                                    graphUtils.stretchCurve(c, orx, ory, nrx, ory, c.maxX, (c.minY + c.maxY)/2);
                                     break;
                                 }
                                 case 7: {
-                                    stretchCurve(c, orx, ory, nrx, ory, c.minX, (c.minY + c.maxY)/2);
+                                    graphUtils.stretchCurve(c, orx, ory, nrx, ory, c.minX, (c.minY + c.maxY)/2);
                                     break;
                                 }
                             }
@@ -1030,12 +1078,12 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             movedSymbol.y += dy;
 
                             reDraw();
-                            drawSymbol(movedSymbol, MOVE_SYMBOL_COLOR);
+                            graphView.drawSymbol(p, movedSymbol, MOVE_SYMBOL_COLOR);
 
                             function detect(knots) {
                                 for (let j = 0; j < knots.length; j++) {
-                                    if (knots[j].symbol == undefined && graphFunctions.getDist(movedSymbol, knots[j]) < MOUSE_DETECT_RADIUS) {
-                                        drawKnotDetect(knots[j], KNOT_DETECT_COLOR);
+                                    if (knots[j].symbol == undefined && graphUtils.getDist(movedSymbol, knots[j]) < MOUSE_DETECT_RADIUS) {
+                                        graphView.drawKnotDetect(p, knots[j], graphView.KNOT_DETECT_COLOR);
                                         return;
                                     }
                                 }
@@ -1058,12 +1106,12 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
 
                             if (clickedKnot != null) {
                                 let knot = clickedKnot;
-                                if (knot.xSymbol == undefined && graphFunctions.getDist(movedSymbol, graphFunctions.createPoint(knot.x, canvasHeight/2)) < MOUSE_DETECT_RADIUS) {
-                                    drawKnotDetect(graphFunctions.createPoint(knot.x, canvasHeight/2), KNOT_DETECT_COLOR);
+                                if (knot.xSymbol == undefined && graphUtils.getDist(movedSymbol, graphUtils.createPoint(knot.x, canvasHeight/2)) < MOUSE_DETECT_RADIUS) {
+                                    graphView.drawKnotDetect(p, graphUtils.createPoint(knot.x, canvasHeight/2), graphView.KNOT_DETECT_COLOR);
                                     return;
                                 }
-                                if (knot.ySymbol == undefined && graphFunctions.getDist(movedSymbol, graphFunctions.createPoint(canvasWidth/2, knot.y)) < MOUSE_DETECT_RADIUS) {
-                                    drawKnotDetect(graphFunctions.createPoint(canvasWidth/2, knot.y), KNOT_DETECT_COLOR);
+                                if (knot.ySymbol == undefined && graphUtils.getDist(movedSymbol, graphUtils.createPoint(canvasWidth/2, knot.y)) < MOUSE_DETECT_RADIUS) {
+                                    graphView.drawKnotDetect(p, graphUtils.createPoint(canvasWidth/2, knot.y), graphView.KNOT_DETECT_COLOR);
                                     return;
                                 }
                             }
@@ -1074,8 +1122,8 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             if (curves.length < CURVE_LIMIT) {
                                 if (drawMode == "curve") {
                                     p.push();
-                                    p.stroke(CURVE_COLORS[drawnColorIdx]);
-                                    p.strokeWeight(CURVE_STRKWEIGHT);
+                                    p.stroke(graphView.CURVE_COLORS[drawnColorIdx]);
+                                    p.strokeWeight(graphView.CURVE_STRKWEIGHT);
                                     if (drawnPts.length > 0) {
                                         let prev = drawnPts[drawnPts.length - 1];
                                         p.line(prev.x, prev.y, current.x, current.y);
@@ -1087,8 +1135,8 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     reDraw();
 
                                     p.push();
-                                    p.stroke(CURVE_COLORS[drawnColorIdx]);
-                                    p.strokeWeight(CURVE_STRKWEIGHT);
+                                    p.stroke(graphView.CURVE_COLORS[drawnColorIdx]);
+                                    p.strokeWeight(graphView.CURVE_STRKWEIGHT);
                                     p.line(lineStart.x, lineStart.y, current.x, current.y);
                                     p.pop();
 
@@ -1127,7 +1175,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             for (let i = 0; i < curves.length; i++) {
                                 let pts = curves[i].pts;
                                 for (let j = 0; j < pts.length; j++) {
-                                    if (graphFunctions.getDist(pts[j], current) < MOUSE_DETECT_RADIUS) {
+                                    if (graphUtils.getDist(pts[j], current) < MOUSE_DETECT_RADIUS) {
                                         clickedCurveIdx = i;
                                         reDraw();
                                         return;
@@ -1213,7 +1261,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 }
                                 for (let j = 0; j < knots.length; j++) {
                                     let knot = knots[j];
-                                    if (knot.symbol == undefined && graphFunctions.getDist(movedSymbol, knot) < MOUSE_DETECT_RADIUS) {
+                                    if (knot.symbol == undefined && graphUtils.getDist(movedSymbol, knot) < MOUSE_DETECT_RADIUS) {
                                         movedSymbol.x = knot.x;
                                         movedSymbol.y = knot.y;
                                         knot.symbol = movedSymbol;
@@ -1242,12 +1290,12 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
 
                             if (!found && clickedKnot != null) {
                                 let knot = clickedKnot;
-                                if (knot.xSymbol == undefined && graphFunctions.getDist(movedSymbol, graphFunctions.createPoint(knot.x, canvasHeight/2)) < MOUSE_DETECT_RADIUS) {
+                                if (knot.xSymbol == undefined && graphUtils.getDist(movedSymbol, graphUtils.createPoint(knot.x, canvasHeight/2)) < MOUSE_DETECT_RADIUS) {
                                     movedSymbol.x = knot.x;
                                     movedSymbol.y = canvasHeight/2;
                                     knot.xSymbol = movedSymbol;
                                     found = true;
-                                } else if (knot.ySymbol == undefined && graphFunctions.getDist(movedSymbol, graphFunctions.createPoint(canvasWidth/2, knot.y)) < MOUSE_DETECT_RADIUS) {
+                                } else if (knot.ySymbol == undefined && graphUtils.getDist(movedSymbol, graphUtils.createPoint(canvasWidth/2, knot.y)) < MOUSE_DETECT_RADIUS) {
                                     movedSymbol.x = canvasWidth/2;
                                     movedSymbol.y = knot.y;
                                     knot.ySymbol = movedSymbol;
@@ -1268,7 +1316,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
 
                                 if (drawMode == "curve") {
                                      // reject if curve drawn is too short
-                                    if (lineSampler.sample(drawnPts).length < 3) {
+                                    if (graphUtils.sample(drawnPts).length < 3) {
                                         return;
                                     }
 
@@ -1291,7 +1339,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     }
 
                                     // sampler.sample, bezier.genericBezier
-                                    let pts = scope.selectedLineType.lineStyle(lineSampler.sample(drawnPts));
+                                    let pts = scope.selectedLineType.lineStyle(graphUtils.sample(drawnPts));
                                     curve = {};
                                     curve.pts = pts;
 
@@ -1310,11 +1358,11 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     curve.minY = minY;
                                     curve.maxY = maxY;
 
-                                    curve.endPt = findEndPts(pts);
-                                    curve.interX = findInterceptX(pts);
-                                    curve.interY = findInterceptY(pts);
-                                    curve.maxima = findTurnPts(pts, 'maxima');
-                                    curve.minima = findTurnPts(pts, 'minima');
+                                    curve.endPt = graphUtils.findEndPts(pts);
+                                    curve.interX = graphUtils.findInterceptX(canvasHeight, pts);
+                                    curve.interY = graphUtils.findInterceptY(canvasWidth, pts);
+                                    curve.maxima = graphUtils.findTurnPts(pts, 'maxima');
+                                    curve.minima = graphUtils.findTurnPts(pts, 'minima');
                                     curve.colorIdx = drawnColorIdx;
 
                                 } else {
@@ -1332,7 +1380,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     for (let i = 0; i <= n; i++) {
                                         let x = lineStart.x + i * sx;
                                         let y = lineStart.y + i * sy;
-                                        pts.push(graphFunctions.createPoint(x, y, i));
+                                        pts.push(graphUtils.createPoint(x, y, i));
                                     }
 
                                     curve = {};
@@ -1343,9 +1391,9 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                     curve.minY = Math.min(lineStart.y, lineEnd.y);
                                     curve.maxY = Math.max(lineStart.y, lineEnd.y);
 
-                                    curve.endPt = findEndPts(pts);
-                                    curve.interX = findInterceptX(pts);
-                                    curve.interY = findInterceptY(pts);
+                                    curve.endPt = graphUtils.findEndPts(pts);
+                                    curve.interX = graphUtils.findInterceptX(canvasHeight, pts);
+                                    curve.interY = graphUtils.findInterceptY(canvasWidth, pts);
                                     curve.maxima = [];
                                     curve.minima = [];
                                     curve.colorIdx = drawnColorIdx;
@@ -1357,7 +1405,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                         let knot = knots[i];
                                         for (let j = 0; j < freeSymbols.length; j++) {
                                             let sym = freeSymbols[j];
-                                            if (graphFunctions.getDist(knot, sym) < 20) {
+                                            if (graphUtils.getDist(knot, sym) < 20) {
                                                 sym.x = knot.x;
                                                 sym.y = knot.y;
                                                 knot.symbol = sym;
@@ -1444,7 +1492,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                     function getMousePt(e) {
                         let x = (e.clientX - 5);
                         let y = (e.clientY - 5);
-                        return (graphFunctions.createPoint(x, y));
+                        return (graphUtils.createPoint(x, y));
                     }
 
                     function isOverSymbol(pt, symbol) {
@@ -1494,645 +1542,15 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
 
                     reDraw = function() {
                         if (curves.length < 4) {
-                            drawBackground();
-                            drawCurves(curves);
+                            graphView.drawBackground(p);
+                            graphView.drawCurves(p, curves);
+                            graphView.drawSymbols(p, freeSymbols);
+                            graphView.drawStretchBox(p, clickedCurveIdx);
                             scope.dat = encodeData();
-                            drawSymbols(freeSymbols);
-                            drawStretchBox(clickedCurveIdx);
                         }
-                    }
+                    };
 
-                    function drawBackground() {
-
-                        function drawHorizontalAxis() {
-                            p.push();
-
-                            p.strokeWeight(CURVE_STRKWEIGHT);
-                            p.strokeJoin(p.ROUND);
-                            p.stroke(0);
-                            p.noFill();
-
-                            let leftMargin = PADDING;
-                            let rightMargin = canvasWidth - PADDING;
-
-                            p.beginShape();
-                            p.vertex(leftMargin, canvasHeight/2);
-                            p.vertex(rightMargin, canvasHeight / 2);
-                            p.vertex(rightMargin - 10, canvasHeight / 2 - 5);
-                            p.vertex(rightMargin, canvasHeight / 2);
-                            p.vertex(rightMargin - 10, canvasHeight / 2 + 5);
-                            p.endShape();
-
-                            p.pop();
-                        }
-
-                        function drawVerticalAxis() {
-                            p.push();
-
-                            p.strokeWeight(CURVE_STRKWEIGHT);
-                            p.strokeJoin(p.ROUND);
-                            p.stroke(0);
-                            p.noFill();
-
-                            let upMargin = PADDING;
-                            let bottomMargin = canvasHeight - PADDING;
-
-                            p.beginShape();
-                            p.vertex(canvasWidth/2, bottomMargin);
-                            p.vertex(canvasWidth/2, upMargin);
-                            p.vertex(canvasWidth/2 - 5, upMargin + 10);
-                            p.vertex(canvasWidth/2, upMargin);
-                            p.vertex(canvasWidth/2 + 5, upMargin + 10);
-                            p.endShape();
-
-                            p.pop();
-                        }
-
-                        function drawGrid() {
-                            p.push();
-
-                            p.noFill();
-                            p.strokeWeight(CURVE_STRKWEIGHT);
-                            p.strokeJoin(p.ROUND);
-                            p.stroke(240);
-
-                            p.push();
-                            p.translate(0, canvasHeight / 2);
-                            let num = canvasHeight / (GRID_WIDTH * 2);
-                            for (let i = 0; i < num; i++) {
-                                p.line(0, -i*GRID_WIDTH, canvasWidth, -i*GRID_WIDTH);
-                                p.line(0, i*GRID_WIDTH, canvasWidth, i*GRID_WIDTH);
-                            }
-                            p.pop();
-
-                            p.push();
-                            p.translate(canvasWidth / 2, 0);
-                            num = canvasWidth / (GRID_WIDTH * 2);
-                            for (let i = 0; i < num; i++) {
-                                p.line(-i*GRID_WIDTH, 0, -i*GRID_WIDTH, canvasHeight);
-                                p.line(i*GRID_WIDTH, 0, i*GRID_WIDTH, canvasHeight);
-                            }
-                            p.pop();
-
-                            p.pop();
-                        }
-
-                        function drawLabel() {
-                            p.push();
-
-                            p.textSize(16);
-                            p.stroke(0);
-                            p.strokeWeight(0.5);
-                            p.fill(0);
-
-                            p.text("O", canvasWidth/2 - 15, canvasHeight/2 + 15);
-                            p.text("x", canvasWidth - PADDING, canvasHeight/2 + 15);
-                            p.text("y", canvasWidth/2 + 5, PADDING);
-
-                            p.pop();
-                        }
-
-                        p.clear();
-                        p.background(255);
-
-                        drawGrid();
-                        drawHorizontalAxis();
-                        drawVerticalAxis();
-                        drawLabel();
-                    }
-
-                    // given a set of points, draw the corresponding curve.
-                    function drawCurve(curve, color) {
-                        if (color == undefined) {
-                            color = CURVE_COLORS[curve.colorIdx];
-                        }
-
-                        p.push();
-                        p.stroke(color);
-                        p.strokeWeight(CURVE_STRKWEIGHT);
-
-
-                        // want to connect closest points x,y wise, not just x wise
-                        let pts = curve.pts;
-                        for (let i = 1; i < pts.length; i++) {
-                            if (pts[i].x - pts[i-1].x < 100 && pts[i].y - pts[i-1].y < 100) {
-                                p.line(pts[i-1].x, pts[i-1].y, pts[i].x, pts[i].y);
-                            }
-                        }
-
-                        // to here for point connection
-
-                        p.pop();
-
-                        curve.endPt = findEndPts(curve.pts);
-                        // draw x intercepts, y intercepts and turning points
-                        drawKnots(curve['interX']);
-                        drawKnots(curve['interY']);
-                        drawKnots2(curve['maxima']);
-                        drawKnots2(curve['minima']);
-                    }
-
-                    function findEndPts(pts) {
-                        if (pts.length == 0) return [];
-
-                        let ends = [];
-
-                        ends.push(graphFunctions.createPoint(pts[0].x, pts[0].y, pts[0].ind));
-                        ends.push(graphFunctions.createPoint(pts[pts.length - 2].x, pts[pts.length - 2].y, pts[pts.length - 2].ind));
-
-                        for (let i = 1; i < pts.length; i++) {
-                            if (pts[i-1].x - pts[i].x > 200) {
-                                ends.push(graphFunctions.createPoint(pts[i-1].x, pts[i-1].y, pts[i-1].ind));
-                                ends.push(graphFunctions.createPoint(pts[i].x, pts[i].y, pts[i].ind));
-                                continue;
-                            }
-                        }
-
-                        if (ends.length == 2) {
-                            for (let i = pts.length - 2; i > 1; i--) {
-                                if (pts[i+1].x - pts[i].x > 200) {
-                                    ends.push(graphFunctions.createPoint(pts[i+1].x, pts[i+1].y, pts[i+1].ind));
-                                    ends.push(graphFunctions.createPoint(pts[i].x, pts[i].y, pts[i].ind));
-                                    continue;
-                                }
-                            }
-                        }
-
-                        return ends;
-                    }
-
-                    function drawCurves(curves, color) {
-                        for (let i = 0; i < curves.length; i++) {
-                            drawCurve(curves[i], color);
-                        }
-                    }
-
-                    // given a set of points, draw the corresponding points (knots).
-                    function drawKnot(knot, color) {
-                        if (color == undefined) {
-                            color = KNOT_COLOR;
-                        }
-
-                        if (knot.symbol != undefined) {
-                            drawSymbol(knot.symbol);
-                        } else {
-                            p.push();
-                            p.noFill();
-                            p.stroke(color);
-                            p.strokeWeight(1.5);
-                            p.line(knot.x - 3, knot.y - 3, knot.x + 3, knot.y + 3);
-                            p.line(knot.x + 3, knot.y - 3, knot.x - 3, knot.y + 3);
-                            p.pop();
-                        }
-                    }
-
-                    function drawKnots(knots, color) {
-                        for (let i = 0; i < knots.length; i++) {
-                            drawKnot(knots[i], color);
-                        }
-                    }
-
-                    function drawKnot2(knot) {
-                        drawKnot(knot);
-
-                        if (knot.xSymbol != undefined) {
-                            drawVerticalDotLine(knot.x, knot.y, canvasHeight/2);
-                            drawSymbol(knot.xSymbol);
-                        }
-
-                        if (knot.ySymbol != undefined) {
-                            drawHorizontalDotLine(knot.y, knot.x, canvasWidth/2);
-                            drawSymbol(knot.ySymbol);
-                        }
-                    }
-
-                    function drawKnots2(knots) {
-                        for (let i = 0; i < knots.length; i++) {
-                            drawKnot2(knots[i]);
-                        }
-                    }
-
-                    function drawKnotDetect(knot) {
-                        p.push();
-                        p.noFill();
-                        p.stroke(KNOT_DETECT_COLOR);
-                        p.strokeWeight(2);
-                        p.line(knot.x - 5, knot.y - 5, knot.x + 5, knot.y + 5);
-                        p.line(knot.x + 5, knot.y - 5, knot.x - 5, knot.y + 5);
-                        p.pop();
-                    }
-
-                    // draw symbols, e.g. "A", "B".
-                    function drawSymbol(symbol, color) {
-                        if (color == undefined) {
-                            color = KNOT_COLOR;
-                        }
-
-                        p.push();
-
-                        p.stroke(color);
-                        p.strokeWeight(1.5);
-                        p.noFill();
-                        p.line(symbol.x - 3, symbol.y - 3, symbol.x + 3, symbol.y + 3);
-                        p.line(symbol.x + 3, symbol.y - 3, symbol.x - 3, symbol.y + 3);
-
-                        p.stroke(0);
-                        p.strokeWeight(0.5);
-                        p.fill(0);
-                        p.textSize(16);
-                        p.text(symbol.text, symbol.x - 5, symbol.y + 20);
-
-                        p.pop();
-                    }
-
-                    function drawSymbols(symbols, color) {
-                        for (let i = 0; i < symbols.length; i++) {
-                            drawSymbol(symbols[i], color);
-                        }
-                    }
-
-                    function drawVerticalDotLine(x, begin, end) {
-                        if (x < 0 || x > canvasWidth) {
-                            return;
-                        }
-
-                        if (begin > end) {
-                            let tmp = begin;
-                            begin = end;
-                            end = tmp;
-                        }
-
-                        p.push();
-                        p.stroke(DOT_LINE_COLOR);
-                        p.strokeWeight(CURVE_STRKWEIGHT);
-
-                        let step = DOT_LINE_STEP;
-                        let toDraw = true;
-                        let y = begin;
-                        while (y + step < end) {
-                            if (toDraw) {
-                                p.line(x, y, x, y+step);
-                            }
-                            y += step;
-                            toDraw = !toDraw;
-                        }
-                        if (toDraw) {
-                            p.line(x, y, x, end);
-                        }
-
-                        p.pop();
-                    }
-
-                    function drawHorizontalDotLine(y, begin, end) {
-                        if (y < 0 || y > canvasHeight) {
-                            return;
-                        }
-
-                        if (begin > end) {
-                            let tmp = begin;
-                            begin = end;
-                            end = tmp;
-                        }
-
-                        p.push();
-                        p.stroke(DOT_LINE_COLOR);
-                        p.strokeWeight(CURVE_STRKWEIGHT);
-
-                        let step = DOT_LINE_STEP;
-                        let toDraw = true;
-                        let x = begin;
-                        while (x + step < end) {
-                            if (toDraw) {
-                                p.line(x, y, x+step, y);
-                            }
-                            x += step;
-                            toDraw = !toDraw;
-                        }
-                        if (toDraw) {
-                            p.line(x, y, end, y);
-                        }
-
-                        p.pop();
-                    }
-
-                    function drawStretchBox(idx) {
-                        if (idx == undefined) {
-                            return;
-                        }
-
-                        let curve = curves[idx];
-
-                        let pts = curve.pts;
-
-                        let minX = curve.minX;
-                        let maxX = curve.maxX;
-                        let minY = curve.minY;
-                        let maxY = curve.maxY;
-
-                        p.push();
-                        p.stroke(DOT_LINE_COLOR);
-                        p.strokeWeight(0.5);
-                        p.line(minX, minY, maxX, minY);
-                        p.line(maxX, minY, maxX, maxY);
-                        p.line(maxX, maxY, minX, maxY);
-                        p.line(minX, maxY, minX, minY);
-
-                        p.fill(255);
-                        p.rect(minX - 4, minY - 4, 8, 8);
-                        p.rect(maxX - 4, minY - 4, 8, 8);
-                        p.rect(minX - 4, maxY - 4, 8, 8);
-                        p.rect(maxX - 4, maxY - 4, 8, 8);
-                        p.triangle((minX + maxX)/2 - 5, minY - 2, (minX + maxX)/2 + 5, minY - 2, (minX + maxX)/2, minY - 7);
-                        p.triangle((minX + maxX)/2 - 5, maxY + 2, (minX + maxX)/2 + 5, maxY + 2, (minX + maxX)/2, maxY + 7);
-                        p.triangle(minX - 2, (minY + maxY) / 2 - 5, minX - 2, (minY + maxY) / 2 + 5, minX - 7, (minY + maxY) / 2);
-                        p.triangle(maxX + 2, (minY + maxY) / 2 - 5, maxX + 2, (minY + maxY) / 2 + 5, maxX + 7, (minY + maxY) / 2); 
-                        p.pop();
-                    }
-
-                    function findInterceptX(pts) {
-                        if (pts.length == 0) return [];
-
-                        let intercepts = [];
-
-                        if (pts[0].y == canvasHeight/2) intercepts.push(pts[0]);
-                        for (let i = 1; i < pts.length; i++) {
-                            if (pts[i].y == canvasHeight/2) {
-                                intercepts.push(graphFunctions.createPoint(pts[i].x, pts[i].y));
-                                continue;
-                            }
-
-                            if ((pts[i-1].y - canvasHeight/2) * (pts[i].y - canvasHeight/2) < 0 && (pts[i-1].y - pts[i].y < Math.abs(200))) {
-                                let dx = pts[i].x - pts[i-1].x;
-                                let dy = pts[i].y - pts[i-1].y;
-                                let grad = dy/dx;
-                                let esti = pts[i-1].x + (1 / grad) * (canvasHeight/2 - pts[i-1].y);
-                                intercepts.push(graphFunctions.createPoint(esti, canvasHeight/2));
-                            }
-                        }
-
-                        return intercepts;
-                    }
-
-                    function findInterceptY(pts) {
-                        if (pts.length == 0) return [];
-
-                        let intercepts = [];
-
-                        if (pts[0].x == canvasWidth/2) intercepts.push(pts[0]);
-                        for (let i = 1; i < pts.length; i++) {
-                            if (pts[i].x == canvasWidth/2) {
-                                intercepts.push(graphFunctions.createPoint(pts[i].x, pts[i].y));
-                                continue;
-                            }
-
-                            if ((pts[i-1].x - canvasWidth/2) * (pts[i].x - canvasWidth/2) < 0 && (pts[i-1].x - pts[i].x < Math.abs(200))) {
-                                let dx = pts[i].x - pts[i-1].x;
-                                let dy = pts[i].y - pts[i-1].y;
-                                let grad = dy/dx;
-                                let esti = pts[i-1].y + grad * (canvasWidth/2 - pts[i-1].x);
-                                intercepts.push(graphFunctions.createPoint(canvasWidth/2, esti));
-                            }
-                        }
-
-                        return intercepts;
-                    }
-
-                    function findTurnPts(pts, mode) {
-                        if (pts.length == 0) {
-                          return [];
-                        }
-
-                        let grad = [];
-                        for (let i = 0; i < pts.length - 1; i++) {
-                            let dx = pts[i+1].x - pts[i].x;
-                            let dy = pts[i+1].y - pts[i].y;
-                            grad.push(dy/dx);
-                        }
-
-                        let turnPts = [];
-
-                        for (let i = 1; i < grad.length; i++) {
-                            if (grad[i-1] != NaN && grad[i] != NaN) {
-                                if (grad[i] * grad[i-1] < 0 && (pts[i].x - pts[i-1].x) * (pts[i+1].x - pts[i].x) > 0) {
-
-                                    let limit = 0.01;
-
-                                    let l = i - 2;
-                                    while (l >= 0 && Math.abs(grad[l]) < limit && Math.abs(grad[l]) > Math.abs(grad[l+1]) && grad[l] * grad[l+1] >= 0) {
-                                        l--;
-                                    }
-                                    if (!(Math.abs(grad[l]) >= limit)) {
-                                        continue;
-                                    }
-
-                                    let r = i + 1;
-                                    while (r < grad.length && Math.abs(grad[r]) < limit && Math.abs(grad[r]) > Math.abs(grad[r-1]) && grad[r] * grad[r-1] >= 0) {
-                                        r++;
-                                    }
-                                    if (!(Math.abs(grad[r]) >= limit)) {
-                                        continue;
-                                    }
-
-                                    let acc1 = grad[l];
-                                    let acc2 = grad[r];
-
-                                    if (mode == 'maxima') {
-                                        if ((pts[i].x > pts[i-1].x && acc1 < 0 && acc2 > 0) || (pts[i].x < pts[i-1].x && acc1 > 0 && acc2 < 0)) {
-                                            turnPts.push(graphFunctions.createPoint(pts[i].x, pts[i].y, pts[i].ind));
-                                        } 
-                                    } else {
-                                        if ((pts[i].x > pts[i-1].x && acc1 > 0 && acc2 < 0) || (pts[i].x < pts[i-1].x && acc1 < 0 && acc2 > 0)) {
-                                            turnPts.push(graphFunctions.createPoint(pts[i].x, pts[i].y, pts[i].ind));
-                                        } 
-                                    }
-
-
-                                }
-                            }
-                        }
-
-                        return turnPts;
-                    }
-
-                    // given a curve, translate the curve
-                    function transCurve(curve, dx, dy) {
-                        let pts = curve.pts;
-
-                        curve.minX += dx;
-                        curve.maxX += dx;
-                        curve.minY += dy;
-                        curve.maxY += dy;
-
-                        for (let i = 0; i < pts.length; i++) {
-                            pts[i].x += dx;
-                            pts[i].y += dy;
-                        }
-
-                        function moveTurnPts(knots) {
-                            for (let i = 0; i < knots.length; i++) {
-                                let knot = knots[i];
-
-                                knot.x += dx;
-                                knot.y += dy;
-
-                                if (knot.symbol != undefined) {
-                                    knot.symbol.x += dx;
-                                    knot.symbol.y += dy;
-                                }
-
-                                if (knot.xSymbol != undefined) {
-                                    knot.xSymbol.x = knot.x;
-                                }
-
-                                if (knot.ySymbol != undefined) {
-                                    knot.ySymbol.y = knot.y;
-                                }
-                            }
-                        }
-
-                        let maxima = curve.maxima;
-                        moveTurnPts(maxima);
-
-                        let minima = curve.minima;
-                        moveTurnPts(minima);
-
-
-                        function moveInter(inter, newInter) {
-                            for (let i = 0; i < inter.length; i++) {
-                                if (inter[i].symbol != undefined) {
-                                    let symbol = inter[i].symbol;
-
-                                    let found = false,
-                                        min = 50,
-                                        knot;
-                                    for (let j = 0; j < newInter.length; j++) {
-                                        if (graphFunctions.getDist(inter[i], newInter[j]) < min) {
-                                            min = graphFunctions.getDist(inter[i], newInter[j]);
-                                            knot = newInter[j];
-                                            found = true;
-                                        }
-                                    }
-
-                                    if (found) {
-                                        symbol.x = knot.x;
-                                        symbol.y = knot.y;
-                                        knot.symbol = symbol;
-                                    } else {
-                                        freeSymbols.push(symbol);
-                                    }
-                                }
-                            }
-                            return newInter;
-                        }
-
-                        let interX = curve.interX,
-                            newInterX = findInterceptX(pts);
-                        curve.interX = moveInter(interX, newInterX);
-
-                        let endPt = curve.endPt,
-                            newEndPt = findEndPts(pts);
-                        curve.endPt = newEndPt;
-
-
-                        let interY = curve.interY,
-                            newInterY = findInterceptY(pts);
-                        curve.interY = moveInter(interY, newInterY);
-
-                        return;
-                    }
-
-                    function stretchCurve(c, orx, ory, nrx, nry, baseX, baseY) {
-
-                        function stretch(pt) {
-                            let nx = (pt.x - baseX) / orx;
-                            let ny = (pt.y - baseY) / ory;
-                            pt.x = nx * nrx + baseX;
-                            pt.y = ny * nry + baseY;
-                        }
-
-                        let pts = c.pts;
-                        for (let j = 0; j < pts.length; j++) {
-                            stretch(pts[j]);
-                            c.pts[j] = pts[j];
-                        }
-
-
-                        function loop1(knots) {
-                            if (knots != undefined) {
-                                for (let j = 0; j < knots.length; j++) {
-                                    let knot = knots[j];
-
-                                    stretch(knot);
-
-                                    if (knot.symbol != undefined) {
-                                        stretch(knot.symbol);
-                                    }
-
-                                    if (knot.xSymbol != undefined) {
-                                        stretch(knot.xSymbol);
-                                    }
-
-                                    if (knot.ySymbol != undefined) {
-                                        stretch(knot.ySymbol);
-                                    }
-                                }
-                            }
-                        }
-
-                        c.endPt = findEndPts(pts)
-
-                        let maxima = c.maxima;
-                        loop1(maxima);
-
-                        let minima = c.minima;
-                        loop1(minima);
-
-                        function loop2(inter, newInter) {
-                            if (inter != undefined) {
-                                for (let i = 0; i < inter.length; i++) {
-                                    if (inter[i].symbol != undefined) {
-                                        let symbol = inter[i].symbol;
-
-                                        let found = false,
-                                            min = 50,
-                                            knot;
-                                        for (let j = 0; j < newInter.length; j++) {
-                                            if (graphFunctions.getDist(inter[i], newInter[j]) < min) {
-                                                min = graphFunctions.getDist(inter[i], newInter[j]);
-                                                knot = newInter[j];
-                                                found = true;
-                                            }
-                                        }
-
-                                        if (found) {
-                                            symbol.x = knot.x;
-                                            symbol.y = knot.y;
-                                            knot.symbol = symbol;
-                                        } else {
-                                            freeSymbols.push(symbol);
-                                        }
-                                    }
-                                }
-                                return newInter;
-                            }
-                        }
-
-                        let interX = c.interX,
-                            newInterX = findInterceptX(pts);
-                        c.interX = loop2(interX, newInterX);
-
-
-                        let interY = c.interY,
-                            newInterY = findInterceptY(pts);
-                        c.interY = loop2(interY, newInterY);
-                    }
-
-                    function clone(obj) {
-                        let json = JSON.stringify(obj);
-                        return JSON.parse(json);
-                    }
-
-                    function encodeData(trunc) {
+                    encodeData = function(trunc) {
 
                         if (trunc == undefined) {
                             trunc = true;
@@ -2152,7 +1570,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                         data.canvasWidth = canvasWidth;
                         data.canvasHeight = canvasHeight;
 
-                        let clonedCurves = clone(curves);
+                        let clonedCurves = graphUtils.clone(curves);
 
                         // sort segments according to their left most points.
                         function compare(curve1, curve2) {
@@ -2170,6 +1588,7 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             else if (min1 == min2) return 0
                             else return 1;
                         }
+
                         clonedCurves.sort(compare);
 
                         function normalise(pt) {
@@ -2182,7 +1601,6 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                                 pt.x = x;
                                 pt.y = y;
                             }
-                            
                         }
 
                         function normalise1(knots) {
@@ -2243,11 +1661,9 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                             normalise2(minima);
                         }
 
-
-
                         data.curves = clonedCurves;
 
-                        let clonedFreeSymbols = clone(freeSymbols);
+                        let clonedFreeSymbols = graphUtils.clone(freeSymbols);
                         for (let i = 0; i < clonedFreeSymbols.length; i++) {
                             let symbol = clonedFreeSymbols[i];
                             normalise(symbol);
@@ -2255,9 +1671,9 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
                         data.freeSymbols = clonedFreeSymbols;
 
                         return data;
-                    }
+                    };
 
-                    function decodeData(rawData) {
+                    decodeData = function(rawData) {
 
                         let data = rawData;
 
@@ -2324,82 +1740,19 @@ define(["p5", "../../../lib/graph_sketcher/bezier.js", "../../../lib/graph_sketc
 
                         reDraw();
                         return;
-                    }
-
-                    scope.undo = function() {
-                        if (p.checkPointsUndo.length == 0 || p.checkPointsUndo === undefined) {
-                            return;
-                        }
-
-                        let checkPointRedo = {};
-                        checkPointRedo.freeSymbolsJSON = JSON.stringify(freeSymbols);
-                        checkPointRedo.curvesJSON = JSON.stringify(curves);
-                        p.checkPointsRedo.push(checkPointRedo);
-
-                        let checkPointUndo = p.checkPointsUndo.pop();
-                        freeSymbols = JSON.parse(checkPointUndo.freeSymbolsJSON);
-                        curves = JSON.parse(checkPointUndo.curvesJSON);
-                        
-                        clickedKnot = null;
-                        clickedCurveIdx = undefined;
-
-                        reDraw();
-                    }
-
-                    scope.redo = function() {
-                        event.stopPropagation();
-                        if (p.checkPointsRedo.length == 0) {
-                            return;
-                        }
-
-                        let checkPointUndo = {};
-                        checkPointUndo.freeSymbolsJSON = JSON.stringify(freeSymbols);
-                        checkPointUndo.curvesJSON = JSON.stringify(curves);
-                        p.checkPointsUndo.push(checkPointUndo);
-
-                        let checkPointRedo = p.checkPointsRedo.pop();
-                        freeSymbols = JSON.parse(checkPointRedo.freeSymbolsJSON);
-                        curves = JSON.parse(checkPointRedo.curvesJSON);
-                        
-                        clickedKnot = null;
-                        clickedCurveIdx = undefined;
-                        reDraw();
-                    }
-
-                    scope.isUndoable = function() {
-                        return (p.checkPointsUndo.length > 0);
-                    }
-
-                    scope.isRedoable = function() {
-                        return (p.checkPointsRedo.length > 0);
-                    }
-
-                    scope.straight = function() {
-                        scope.selectedLineType = linearLineType;
-                    }
-
-                    scope.poly = function() {
-                        scope.selectedLineType = bezierLineType;
-                    }
-
-                    scope.clean = function() {
-                        p.checkPoint = {};
-                        p.checkPoint.freeSymbolsJSON = JSON.stringify(freeSymbols);
-                        p.checkPoint.curvesJSON = JSON.stringify(curves);
-                        p.checkPointsUndo.push(p.checkPoint);
-                        p.checkPointsRedo = [];
-
-                        curves = [];
-                        clickedKnot = null;
-                        clickedCurveIdx = undefined;
-                        initiateFreeSymbols();
-                        reDraw();
-                    }
-
-                    scope.submit = function() {
-                        $("#graphModal").foundation("reveal", "close");
                     };
                 }
+
+                scope.logOnClose = function(event) {
+                    // This ought to catch people who navigate away without closing the editor!
+                    if (scope.log != null) {
+                        scope.log.actions.push({
+                            event: "NAVIGATE_AWAY",
+                            timestamp: Date.now()
+                        });
+                        api.logger.log(scope.log);
+                    }
+                };
 
                 $rootScope.showGraphSketcher = function(initialState, questionDoc, editorMode) {
                     return new Promise(function(resolve, reject) {
